bertelastic_git_2/example/requirements.txt
bert-serving-client==1.10.0
elasticsearch==7.0.4
pandas==0.25.1
gensim
